backend/pyproject.toml
[project]
name = "marga-manager-backend"
version = "0.1.0"
description = "Marga Manager — SaaS система управления кухней и инвентарём"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.4",
    "django-filter>=24.3",
    "drf-spectacular>=0.27",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "gunicorn>=22.0",
    "django-unfold>=0.40",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
